Fix namespaces for moved scope classes
#563
